refactor: 공용 프로필 컴포넌트 정리 및 폴더 컨벤션 동기화
