+++ src/pages/my-info/my-info-profile.tsx
@@ -5,7 +5,7 @@ import { projectQueries } from '@apis/project/project-queries';
 import { reportQueries } from '@apis/report/report-queries';
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 import { useMemo, useState } from 'react';
-import ProfileDetail from '../../shared/templates/profileDetail';
+import ProfileDetail from '@components/profile/ProfileDetail';
 
 const MyInfoProfile = () => {
   const [year, setYear] = useState(new Date().getFullYear());
