+++ src/shared/components/profile/ProfileDetail.tsx
